Unified cop AI code structure, migrated to shared NPC behaviors, and introduced `CopDeathEvent`.

+++ gangland-features/cops-n-crooks/src/main/java/me/luckyraven/copsncrooks/police/config/CopSettings.java
@@ -107,7 +107,7 @@ public interface CopSettings {
 
 	double getRangedMaxDistance();
 
-	double getMinRepathAfterLossTicks();
+	int getMinRepathAfterLossTicks();
 
 	int getMaxReturnTicks();
